+++ src/templates/node-express-clean-ts/src/server.ts
@@ -1,7 +1,0 @@
-import { app } from "./app";
-
-const port = process.env.PORT || 3333;
-
-app.listen(port, () =>
-  console.log(`🎉 Server started on http://localhost:${port} 🎉`)
-);