+++ src/templates/node-express-js/config/index.js
@@ -1,9 +1,0 @@
-const { NODE_PORT, NODE_URL } = process.env;
-
-const port = NODE_PORT;
-const baseUrl = NODE_URL;
-
-module.exports = {
-  port,
-  baseUrl,
-};