+++ src/templates/node-express-js/src/errors/notFound.js
@@ -1,9 +1,0 @@
-const notFound = (req, res, next) => {
-  const err = new Error('Not Found');
-
-  err.status = 404;
-
-  next(err);
-};
-
-module.exports = notFound;