+++ src/templates/node-express-js/src/index.js
@@ -1,7 +1,0 @@
-require('../config/env');
-const server = require('./server');
-const { port, baseUrl } = require('../config');
-
-server.listen(port, () =>
-  console.log(`Server started on ${baseUrl}:${port} ✨`)
-);