+++ src/templates/node-express-js/src/routes/index.js
@@ -1,9 +1,0 @@
-const { Router } = require('express');
-
-const routes = Router();
-
-routes.get('/', (_req, res) =>
-  res.json({ message: 'Welcome to Generator NodeJS' })
-);
-
-module.exports = routes;