+++ src/templates/node-express-ts/prettier.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  // tabWidth: 2,
-  semi: true,
-  singleQuote: true,
-  arrowParens: 'avoid',
-  // trailingComma: 'es5',
-};