+++ src/templates/node-express-ts/src/@types/process.d.ts
@@ -1,7 +1,0 @@
-declare namespace NodeJS {
-  export interface ProcessEnv {
-    NODE_ENV: 'development' | 'production' | 'test';
-    PORT?: string;
-    URL: string;
-  }
-}