+++ src/templates/node-express-ts/src/errors/index.ts
@@ -1,3 +1,0 @@
-export * from './HandleErrorNotFound';
-export * from './HandleMiddlewareError';
-export * from './HttpException';