+++ src/templates/node-express-ts/src/routes/index.ts
@@ -1,9 +1,0 @@
-import { Request, Response, Router } from 'express';
-
-const router = Router();
-
-router.get('/', (_req: Request, res: Response) => {
-  return res.json({ message: 'Welcome to Generator NodeJS API' });
-});
-
-export { router };