+++ src/templates/node-fastify-js/src/database/index.js
@@ -1,9 +1,0 @@
-import mongoose from 'mongoose'
-
-mongoose
-  .connect(process.env.MONGODB_URL, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  })
-  .then(() => console.log(`✨ MongoDB connected`))
-  .catch((err) => console.log(`MongoDB connection error: ${err}`))