+++ src/templates/node-fastify-js/src/models/User.js
@@ -1,9 +1,0 @@
-import { Schema, model } from 'mongoose'
-
-const UserSchema = new Schema({
-  username: String,
-  password: String,
-  email: String,
-})
-
-export default model('User', UserSchema)